requirements.txt
gymnasium==0.29.1
numpy==1.26.4
matplotlib==3.8.3
pandas==2.2.1
torch==2.2.1
stable-baselines3==2.2.1
fastapi==0.110.0
uvicorn==0.27.1
openai>=1.0.0
python-dotenv>=1.0.0